requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Data ingestion
ccxt
requests
web3
pydantic
pydantic-settings
pycoingecko
ta-lib
praw

# Databases
psycopg2-binary
sqlalchemy
timescale
chromadb
redis
celery
python-dotenv
pyyaml

# ML / Analytics (CPU-only)
tensorflow-cpu==2.20.0
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
scikit-learn
prophet
shap
lime
sentence-transformers
pytorch-forecasting
pandas
numpy>=1.21,<2.0
statsmodels
vaderSentiment
seaborn 
matplotlib

# MLOps
mlflow

# MCP
mcp
mcp-server

# Dev tools
black
isort
flake8
mypy
pytest
pytest-asyncio

